cli: add IPv6 support on the server side
Even though the CLI client is perfectly able to connect to an IPv6
client, there were no way to configure the daemon to listen to an
IPv6 address. This patch fixes that.

As a bonus, it is now possible to use plain hostnames that will get
resolved (such as `localhost`).